validate: CheckLinux is platform dependent
