+++ validate/validate_unsupported.go
@@ -10,3 +10,8 @@ import (
 func LastCap() capability.Cap {
 	return capability.Cap(-1)
 }
+
+// CheckLinux is a noop on this platform
+func (v *Validator) CheckLinux() (errs error) {
+	return nil
+}
